requirements.txt
Django~=4.1.2
djangorestframework
environ

# Django environment variables
# https://django-environ.readthedocs.io/
django-environ

#Postgres Requirement
# sudo apt install python3-dev libpq-dev
#development
psycopg2-binary
#production ready
# psycopg2
